# swaggerConfigJson.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.1",
    "title": "Users Api Test",
    "description": "This api can help you out with the profound understanding of swagger API",
    "contact": {
      "name": "Erfan Motallebi",
      "email": "Erfan.Motallebi1990@gmail.com",
      "url": "http://www.github.com/Erfan-Motallebi"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
    "/api/users": {
      "get": {
        "tags": ["Users"],
        "description": "You can get all the users from the database",
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "age": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["User"],
        "description": "Get a specific user through user ID",
        "operationId": "getUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number"
                  },
                  "name": {
                    "type": "string"
                  },
                  "age": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "age": {
                      "type": "number"
                    }
                  },
                  "example": {
                    "id": 1,
                    "name": "Josh",
                    "age": 54
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/user/{id}": {
      "get": {
        "tags": ["Specific User"],
        "description": "get a Specific User through user ID",
        "operationId": "getUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": {
                "id": 1
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "age": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "id": 1,
                    "name": "Josh",
                    "age": "57"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  }
}